+++ src/ext-test/java/org/opentripplanner/ext/fares/FaresFilterTest.java
@@ -19,12 +19,14 @@ import org.opentripplanner.transit.model.basic.Money;
 
 public class FaresFilterTest implements PlanTestConstants {
 
+  private final TransitModelForTest testModel = TransitModelForTest.of();
+
   @Test
   void shouldAddFare() {
     final int ID = 1;
 
     Itinerary i1 = newItinerary(A, 0)
-      .walk(20, Place.forStop(TransitModelForTest.stopForTest("1:stop", 1d, 1d)))
+      .walk(20, Place.forStop(testModel.stop("1:stop", 1d, 1d).build()))
       .bus(ID, 0, 50, B)
       .bus(ID, 52, 100, C)
       .build();